# backend/app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ChronoSift API",
    "description": "Forensic Timeline Analysis API - Comprehensive documentation for all endpoints",
    "version": "1.0.0",
    "contact": {
      "name": "ChronoSift",
      "url": "https://github.com/yourusername/chronosift"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth/login"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "Timeline": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "name": {"type": "string"},
          "description": {"type": "string"},
          "project_id": {"type": "integer"},
          "is_master": {"type": "boolean"},
          "entry_count": {"type": "integer"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
...
    }
  },
  "paths": {
...
    "/timelines": {
      "post": {
        "tags": ["Timelines"],
        "summary": "Create a new timeline",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "project_id"],
                "properties": {
                  "name": {"type": "string"},
                  "description": {"type": "string"},
                  "project_id": {"type": "integer"},
                  "columns": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {"type": "string"},
                        "column_type": {"type": "string"},
                        "is_required": {"type": "boolean"}
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Timeline created",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Timeline"}
              }
            }
          }
        }
      }
    },
    "/timelines/{id}": {
      "get": {
        "tags": ["Timelines"],
        "summary": "Get timeline details",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          },
          {
            "name": "include_entries",
            "in": "query",
            "schema": {"type": "boolean"}
          }
        ],
        "responses": {
          "200": {
            "description": "Timeline details",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Timeline"}
              }
            }
          }
        }
      }
    },
    "/timelines/{timeline_id}/search": {
      "get": {
        "tags": ["Timelines"],
        "summary": "Search timeline entries with advanced filtering",
        "description": "Search timeline entries using PostgreSQL JSONB search capabilities. Supports text search, field-specific search, exact match filtering, and timestamp range filtering for time pivoting.",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "timeline_id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID of the timeline to search"
          },
          {
            "name": "q",
            "in": "query",
            "schema": {"type": "string"},
            "description": "Text to search for across all fields (full-text search using PostgreSQL tsvector)"
          },
          {
            "name": "field",
            "in": "query",
            "schema": {"type": "string"},
            "description": "Specific field name to search within (used with 'q' parameter). Examples: 'Description', 'Tags', 'Source'"
          },
          {
            "name": "filter",
            "in": "query",
            "schema": {"type": "string"},
            "description": "JSON-encoded filter object for exact matches or timestamp ranges. Supports two formats: 1) Exact match: {\"FieldName\": \"value\"} for JSONB containment matching. 2) Timestamp range: {\"Timestamp\": {\"start\": \"2024-01-01T00:00:00Z\", \"end\": \"2024-01-02T00:00:00Z\"}} for time pivoting.",
            "examples": {
              "exact_match": {
                "value": "{\"Type\": \"Process Creation\", \"User\": \"admin\"}",
                "summary": "Find entries where Type is 'Process Creation' and User is 'admin'"
              },
              "timestamp_range": {
                "value": "{\"Timestamp\": {\"start\": \"2024-02-09T18:47:10.022Z\", \"end\": \"2024-02-09T19:57:10.022Z\"}}",
                "summary": "Find entries within a 1-hour time window (used by Time Pivot feature)"
              },
              "combined": {
                "value": "{\"Timestamp\": {\"start\": \"2024-02-09T18:00:00Z\", \"end\": \"2024-02-09T20:00:00Z\"}, \"Source\": \"Windows Event Log\"}",
                "summary": "Combine timestamp range with exact field match"
              }
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "schema": {"type": "string"},
            "description": "Field name to sort results by. Defaults to entry creation time if not specified."
          },
          {
            "name": "sort_dir",
            "in": "query",
            "schema": {"type": "string", "enum": ["asc", "desc"], "default": "asc"},
            "description": "Sort direction: 'asc' for ascending, 'desc' for descending"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {"type": "integer", "default": 1},
            "description": "Page number for pagination"
          },
          {
            "name": "per_page",
            "in": "query",
            "schema": {"type": "integer", "default": 50},
            "description": "Number of results per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results with pagination metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "entries": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "description": "Timeline entry with JSONB data field"
                      }
                    },
                    "total": {"type": "integer", "description": "Total number of matching entries"},
                    "page": {"type": "integer", "description": "Current page number"},
                    "per_page": {"type": "integer", "description": "Results per page"},
                    "pages": {"type": "integer", "description": "Total number of pages"},
                    "query": {"type": "string", "description": "The search query that was executed"},
                    "filter": {"type": "string", "nullable": true, "description": "The filter that was applied"}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter JSON or search parameters",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"}
              }
            }
          },
          "404": {
            "description": "Timeline not found"
          }
        }
      }
    },
...
  }
}